+++ semester2/1homework/2/stack.h
@@ -4,9 +4,9 @@ class Stack
 {
 public:
     virtual ~Stack() {}
-    virtual void push(double value) = 0;
+    virtual void push(int value) = 0;
     virtual int length() = 0;
-    virtual double pop() = 0;
-    virtual double top() = 0;
+    virtual int pop() = 0;
+    virtual int top() = 0;
     virtual void print() = 0;
 };
